Flush output buffer completely before exiting.
This is only done whne the core.process.exit is set to true.

This will completely flush the response immediately so any
background processes queued with register_shutdown_function
will run in the background and the client will not have to wait for them
to finish to get the response.

(Used when sending email in background )